python-and-cue/pyproject.toml
[project]
name = "pythonandcue"
description = "Python And Cue Example"
authors = [
    {name = "Victor Payno", email = "vpayno@users.noreply.github.com"},
]
maintainers = [
    {name = "Victor Payno", email = "vpayno@users.noreply.github.com"},
]
dependencies = [
    "pycue>=0.1.3",
    "rich>=13.7.1",
]
requires-python = ">=3.12"
dynamic = ["version"]
readme = "README.md"
license = {file = "../LICENSE"}
keywords = ["cue", "python", "yaml"]

[project.urls]
"Homepage" = "https://github.com/vpayno/cuelang-workspace"
"Bug Tracker" = "https://github.com/vpayno/cuelang-workspace/issues"
"Repository" = "https://github.com/vpayno/cuelang-workspace.git"
"Changelog" = "https://github.com/vpayno/cuelang-workspace/releases"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pyright]
include = ["src", "test"]

[tool.ruff]
line-length = 120
indent-width = 4

[tool.ruff.lint]
# select = []
# ignore = []
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"

[tool.mypy]
check_untyped_defs = true
disallow_untyped_calls = true
disallow_untyped_decorators = true
disallow_untyped_defs = true
no_implicit_optional = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
strict_equality =true
warn_return_any = true

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120
skip_gitignore = true

[tool.pdm]
distribution = true
[tool.pdm.dev-dependencies]
dev = [
    "pytest>=8.1.1",
    "pytest-cov>=4.1.0",
    "pytest-randomly>=3.15.0",
    "pytest-datafiles>=3.0.0",
    "mypy>=1.9.0",
    "mypy-extensions>=1.0.0",
    "pyright>=1.1.353",
    "refurb>=2.0.0",
    "ruff>=0.3.2",
    "pylint>=3.1.0",
    "bandit>=1.7.8",
    "mkdocs>=1.5.3",
    "pdoc3>=0.10.0",
]

[tool.pdm.version]
source = "scm"
write_to = "pythonandcue/_version.py"
write_template = "__version__ = '{}'"

[tool.pdm.scripts]
main = { call = "pythonandcue:main" }
module = { shell = "python -m pythonandcue" }

clean_pycache = { shell = "find . -type f -name '*.py[co]' -delete -print -o -type d -name __pycache__ -delete -print" }
clean_build = { shell = "rm -rf dist" }
clean = { composite = ["clean_pycache", "clean_build"] }

format = { shell = "ruff format" }

ruff = { shell = "ruff check" }
pylint = { shell = "pylint --ignore-patterns='_[a-z][a-z0-9_-]+.py' ./src ./test" }
pyright = { shell = "pyright --stats" }
refurb = { shell = "refurb ./src ./test" }
bandit = { shell = "bandit --verbose --recursive ./src" }
mypy = { shell = "mypy --install-types --non-interactive ./src ./test" }
lint = { composite = ["ruff", "pyright", "pylint", "refurb", "mypy"] }

doctest = { shell = "python -m doctest -v ./src/*/[a-z]*.py" }
pytest = { shell = "pytest --verbose --cov=. --cov-branch --cov-report=term-missing --cov-report=xml:.coverage.xml -p no:randomly ./test" }
coverage_report = { shell = "coverage report --show-missing" }
coverage_annotate = { shell = "coverage annotate" }
test = { composite = ["doctest", "pytest"] }

full = { composite = ["clean", "format", "lint", "test"] }

[project.scripts]
pythonandcue = "pythonandcue.main:main"
